withdrawal fix

+++ internal/repository/order_repository.go
@@ -24,8 +24,8 @@ type OrderData struct {
 }
 
 type UserBalance struct {
-	Current   int `json:"current"`
-	Withdrawn int `json:"withdrawn"`
+	Current   float32 `json:"current"`
+	Withdrawn float32 `json:"withdrawn"`
 }
 
 func (or *OrderRepository) IsOrderExist(orderNumber string, userID int) (int, error) {
